streamlit/requirements.txt
python-binance
requests
pandas
numpy
jupyterlab
pyinstaller
binance-connector
sklearn

--find-links https://download.pytorch.org/whl/torch_stable.html

torch==1.9.1+cu111
torchvision

alembic
altair
asttokens
attrs
backcall
blinker
blis
Boruta
cachetools
catalogue
certifi
charset-normalizer
click
cloudpickle
colorama
colorlover
commonmark
cufflinks
cycler
cymem
databricks-cli
debugpy
decorator
docker
entrypoints
executing
Flask
fonttools
funcy
future
gensim
gitdb
GitPython
greenlet
htmlmin
idna
ImageHash
imbalanced-learn
importlib-metadata
ipykernel
ipython
ipywidgets
itsdangerous
jedi
Jinja2
joblib
jsonschema
jupyter_client
jupyter_core
jupyterlab-widgets
kiwisolver
kmodes
lightgbm
llvmlite
Mako
MarkupSafe
matplotlib
matplotlib-inline
missingno
mlflow
mlxtend
multimethod
murmurhash
nest-asyncio
networkx
nltk
numba
numexpr
numpy
oauthlib
packaging
pandas
pandas-profiling
parso
patsy
phik
pickleshare
Pillow
plac
plotly
preshed
prometheus-client
prometheus-flask-exporter
prompt-toolkit
protobuf
psutil
pure-eval
pyarrow
pycaret
pydantic
pydeck
Pygments
PyJWT
pyLDAvis
Pympler
pynndescent
pyod
pyparsing
pyrsistent
python-dateutil
pytz
pytz-deprecation-shim
PyWavelets
pywin32
PyYAML
pyzmq
querystring-parser
regex
requests
rich
scikit-learn
scikit-plot
scipy
seaborn
semver
six
sklearn
smart-open
smmap
spacy
SQLAlchemy
sqlparse
srsly
stack-data
statsmodels
streamlit
streamlit-pandas-profiling
tabulate
tangled-up-in-unicode
tenacity
textblob
thinc
threadpoolctl
toml
toolz
tornado
tqdm
traitlets
typing_extensions
tzdata
tzlocal
umap-learn
urllib3
validators
visions
waitress
wasabi
watchdog
wcwidth
websocket-client
Werkzeug
widgetsnbextension
wordcloud
yellowbrick
zipp
